+++ src/main/java/AILib/layers/StaticLayer.java
@@ -112,7 +112,7 @@ public class StaticLayer implements Layer {
     }
 
     @Override
-    public void trainLayer(double[] outputs, float ratio) {
+    public void trainLayer(double[] outputs, double ratio) {
         for (Neuron neuron : this.neurons) {
             for (int j = 0; j < neuron.weights.size(); j++)
                 neuron.weights.set(j,
